+++ docs/Changelog.md
@@ -78,7 +78,7 @@
 * [#7000](https://github.com/TouK/nussknacker/pull/7000) Show all possible options for dictionary editor on open.
 * [#6979](https://github.com/TouK/nussknacker/pull/6979) Introduces an activities panel that provides information about all system activities.
 * [#7058](https://github.com/TouK/nussknacker/pull/7058) Performance optimization: Add missing Flink TypeInformation for better serialization
-* [#7021](https://github.com/TouK/nussknacker/pull/7021) Definitions service can return definition without config from an additional provider
+* [#7021](https://github.com/TouK/nussknacker/pull/7021) Definitions service can return definition without UI config
 
 ## 1.17
