+++ docs/MigrationGuide.md
@@ -57,7 +57,7 @@ To see the biggest differences please consult the [changelog](Changelog.md).
 * [#7021](https://github.com/TouK/nussknacker/pull/7021)
   * Definitions API changes:
       * GET `/api/processDefinitionData/*}` 
-        * added optional query param `enrichedWithAdditionalConfig`
+        * added optional query param `enrichedWithUiConfig`
         * added `requiredParam` property to the response for parameter config at `components['component-id'].parameters[*]`
 
 ### Configuration changes
